+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2026-04-15]
+## [1.0.0 - 2026-04-16]
 ### Added
 - First version of the project
 - Spring Application
@@ -254,6 +254,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Project context in form template context
 - Template Filename
 - Send a query with different project types to bridgehead
+- List parameters in controller
 
 
 ### Changed